stash: construction tests
