+++ configs/optimism/goerli-erc20.json
@@ -19,14 +19,27 @@
   "all_in_memory_enabled": true,
   "table_size": 1,
   "value_log_file_size": 1024,
-  "construction": null,
+  "construction": {
+    "stale_depth": 3,
+    "broadcast_limit": 5,
+    "ignore_broadcast_failures": false,
+    "clear_broadcasts": true,
+    "constructor_dsl_file": "goerli-erc20.ros",
+    "quiet": false,
+    "initial_balance_fetch_disabled": false,
+    "end_conditions": {
+      "create_account": 10,
+      "transfer": 20
+    }
+  },
   "data": {
     "start_index": 4307133,
-    "reconciliation_disabled": true,
+    "reconciliation_disabled": false,
     "inactive_discrepancy_search_disabled": true,
-    "balance_tracking_disabled": true,
-    "initial_balance_fetch_disabled":true,
+    "balance_tracking_disabled": false,
+    "initial_balance_fetch_disabled": false,
     "active_reconciliation_concurrency": 32,
+    "historical_balance_disabled": false,
     "bootstrap_balances": "",
     "log_balance_changes": true,
     "log_transactions": true,
